Re-merging small files, implementing some of suggested naming feedback
