+++ rosenpass/src/config/mod.rs
@@ -13,20 +13,76 @@ pub use rosenpass_cfg::*;
 mod rosenpass_peer;
 pub use rosenpass_peer::*;
 
-mod rosenpass_keypair;
-pub use rosenpass_keypair::*;
-
-mod protocol_version;
-pub use protocol_version::*;
-
-mod verbosity;
-pub use verbosity::*;
-
-mod wireguard;
-pub use wireguard::*;
-
-mod peer_osk_domain_seperator;
-pub use peer_osk_domain_seperator::*;
-
-mod util;
+pub mod util;
 pub use util::*;
+
+use serde::{Deserialize, Serialize};
+use std::path::Path;
+use std::path::PathBuf;
+
+/// Public key and secret key locations.
+#[derive(Debug, Deserialize, Serialize, PartialEq, Eq, Clone)]
+#[serde(deny_unknown_fields)]
+pub struct RosenpassKeypair {
+    /// path to the public key file
+    pub public_key: PathBuf,
+
+    /// path to the secret key file
+    pub secret_key: PathBuf,
+}
+
+impl RosenpassKeypair {
+    /// Construct a keypair from its fields
+    pub fn new<Pk: AsRef<Path>, Sk: AsRef<Path>>(public_key: Pk, secret_key: Sk) -> Self {
+        let public_key = public_key.as_ref().to_path_buf();
+        let secret_key = secret_key.as_ref().to_path_buf();
+        Self {
+            public_key,
+            secret_key,
+        }
+    }
+}
+
+/// The protocol version to be used by a peer.
+#[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Copy, Clone, Default)]
+#[serde(deny_unknown_fields)]
+pub enum ProtocolVersion {
+    #[default]
+    V02,
+    V03,
+}
+/// Information for supplying exchanged keys directly to WireGuard
+#[derive(Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub struct WireGuard {
+    /// Name of the WireGuard interface to supply with pre-shared keys generated by the Rosenpass
+    /// key exchange
+    pub device: String,
+
+    /// WireGuard public key of the peer to supply with pre-shared keys
+    pub peer: String,
+
+    /// Extra parameters passed to the `wg` command
+    #[serde(default)]
+    pub extra_params: Vec<String>,
+}
+
+/// Level of verbosity for [crate::app_server::AppServer]
+///
+/// The value of the field [crate::app_server::AppServer::verbosity]. See the field documentation
+/// for details.
+///
+/// - TODO: replace this type with [`log::LevelFilter`], also see <https://github.com/rosenpass/rosenpass/pull/246>
+#[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Copy, Clone)]
+#[serde(deny_unknown_fields)]
+pub enum Verbosity {
+    Quiet,
+    Verbose,
+}
+
+impl Default for Verbosity {
+    /// Self::Quiet
+    fn default() -> Self {
+        Self::Quiet
+    }
+}
