--- rosenpass/src/config/wireguard.rs
@@ -1,17 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-/// Information for supplying exchanged keys directly to WireGuard
-#[derive(Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct WireGuard {
-    /// Name of the WireGuard interface to supply with pre-shared keys generated by the Rosenpass
-    /// key exchange
-    pub device: String,
-
-    /// WireGuard public key of the peer to supply with pre-shared keys
-    pub peer: String,
-
-    /// Extra parameters passed to the `wg` command
-    #[serde(default)]
-    pub extra_params: Vec<String>,
-}
